store data order beli sekaarang

+++ database/migrations/2023_12_06_135327_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('tbl_orders', function (Blueprint $table) {
             $table->id('id_orders');
             $table->foreignId('user_id')->references('id')->on('users');
-            $table->integer('no_order')->nullable();
+            $table->bigInteger('no_order')->nullable();
             $table->integer('bayar')->nullable();
             $table->integer('total_bayar')->nullable();
             $table->string('status');
